filings-backend/requirements.txt
anyio==3.6.2
astroid==2.13.2
bcrypt==4.0.1
certifi==2022.12.7
cloudinary==1.30.0
colorama==0.4.6
dill==0.3.6
fastapi==0.89.0
greenlet==2.0.1
idna==3.4
isort==5.11.4
lazy-object-proxy==1.9.0
mccabe==0.7.0
passlib==1.7.4
platformdirs==2.6.2
psycopg2-binary==2.9.5
pydantic==1.10.4
PyJWT==2.6.0
pylint==2.15.10
pypdf2==3.0.1
python-multipart==0.0.5
PyYAML==6.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.46
starlette==0.22.0
tomli==2.0.1
tomlkit==0.11.6
typing-extensions==4.4.0
urllib3==1.26.13
wrapt==1.14.1
